Moved melee weapon function. Moved function to vector2utils
